modules/videomonitoring/requirements-amd64.txt
#azure-iothub-device-client
azure-iot-device~=2.0.0rc12
Flask
numpy
opencv-contrib-python==4.1.1.26
imutils==0.5.3